Editing interface for backend process
